blockmanager: make header list optional for block validation
This commit is a preparation for re-using the block validation code for
the side loading of block headers. Since we fall back to the block
header store if the header list (LRU cache for recent headers) doesn't
have the header we are looking for, we might as well make the lookup
optional if the header list is not specified.