headerfs+mock: reduce number of allocations
This is a memory allocation optimization improvement commit.
We reduce the number of buffers and readers needed to read a block or
filter header by directly reading from the file on disk into the target
struct or array.

Interestingly, for de-serializing a block header, which uses a set of
wire.readElement() calls, reading from the file directly is much, much
slower than reading 80 bytes into memory and then reading the individual
elements from there. The assumption is that this is due to the fact that
each io.ReadFull() call within the readElement() internals causes an
individual disk access.

To illustrate, here are the benchmark results for the slower version
which calls header.Deserialize() directly on the file:

BenchmarkBlockHeaderStoreRead-16     526   2211137 ns/op   22400 B/op     200 allocs/op

As a comparison, these are the results with the committed version that
first reads 80 bytes into a (pooled) buffer, then de-serializes from
that:

BenchmarkBlockHeaderStoreRead-16    2437    512484 ns/op   24818 B/op     300 allocs/op